Extract makeMap and makeSet to utils
