+++ src/utils/collections.js
@@ -1,0 +1,20 @@
+function makeMap(iterable, keyFunc, valueFunc) {
+  const map = {};
+  iterable.forEach(item => {
+    const key = keyFunc(item);
+    const value = valueFunc(item);
+    map[key] = value;
+  });
+  return map;
+}
+
+function makeSet(iterable, func) {
+  const set = new Set();
+  iterable.forEach(item => set.add(func(item)));
+  return set;
+}
+
+module.exports = {
+  makeMap: makeMap,
+  makeSet: makeSet
+};
